Update Installer and Setup
