+++ apps/_docs/content/docs/setup.mdx
@@ -21,10 +21,10 @@ Learn how to [install the Stripe CLI](https://docs.stripe.com/stripe-cli#install
 
 ### Cloning
 
-Clone the repo using `next-forge` init.
+Clone the repo using `next-forge` init, replacing `[my-project]` with your project name.
 
 ```sh title="Terminal"
-npx next-forge init
+npx next-forge init [my-project]
 ```
 
 ### Login
